versioning_function/requirements.txt
google-cloud-storage
Flask